moved sections from pages to separate folder called 'sections'
